Added outline: none; to languages button
I know we are not supposed to make changes on our homework after turning them in but I had to get rid of the annoying border around the language choices button (appeares when you click on it). I just found the solution for that. I hope it is not a problem.
Now I can sleep in peace...